requirements.txt
# Core NLP and AI Libraries
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai==1.6.1

# Vector Database and Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.7

# Data Processing
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.2

# Web Framework (for chatbot interface)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For advanced features
# transformers==4.36.0  # For local model fine-tuning
# torch==2.1.1  # PyTorch for local models
# accelerate==0.25.0  # For model acceleration 